Fix install script so that it asks if you want to reinstall virtualenv and if you want to overwrite the database. Fix recreatedb.sh so that it doesn't constantly prompt for password. Removed the proxying of /mox/ & /*.jsp from tomcat.conf.

+++ oio_rest/install.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
 
-while getopts ":ds" OPT; do
+
+while getopts ":ys" OPT; do
   case $OPT in
-	s)
-		SKIP_SYSTEM_DEPS=1
-		;;
-	*)
-		echo "Usage: $0 [-d] [-s]"
-		echo "	-s: Skip installing oio_rest API system dependencies"
-		exit 1;
-		;;
-	esac
+        s)
+                SKIP_SYSTEM_DEPS=1
+                ;;
+        y)
+                ALWAYS_CONFIRM=1
+                ;;
+        *)
+                echo "Usage: $0 [-y] [-s]"
+                echo "  -s: Skip installing oio_rest API system dependencies"
+                echo "  -y: Always confirm (yes) when prompted"
+                exit 1;
+                ;;
+        esac
 done
 
 # Get the folder of this script
@@ -45,46 +50,73 @@ sudo chown www-data "$MOX_STORAGE"
 # Setup and start virtual environment
 VIRTUALENV="$DIR/python-env"
 
-echo "Setting up virtual enviroment '$VIRTUALENV'"
+CREATE_VIRTUALENV=0
+
 if [ -d $VIRTUALENV ]; then
-	echo "$VIRTUALENV already existed. Removing."
-	rm -rf $VIRTUALENV
-fi
-
-echo "Creating virtual enviroment '$VIRTUALENV'"
-virtualenv $VIRTUALENV
-
-if [ ! -d $VIRTUALENV ]; then
-	echo "Failed creating virtual environment!"
-	exit 1
+	if [ -z $ALWAYS_CONFIRM ]; then
+		echo "$VIRTUALENV already existed."
+		read -p "Do you want to reinstall it? (y/n): " -n 1 -r
+		echo
+		if [[ $REPLY =~ ^[Yy]$ ]]; then
+			CREATE_VIRTUALENV=1
+		else
+			CREATE_VIRTUALENV=0
+		fi
+	else
+		CREATE_VIRTUALENV=1
+	fi
+	if [ $CREATE_VIRTUALENV == 1 ]; then
+		rm -rf $VIRTUALENV
+	fi
 else
-	echo "Virtual environment created. Starting..."
-	source $VIRTUALENV/bin/activate
-
-	pushd "$DIR"
-	python setup.py develop
-	popd
-
-	echo "Stopping virtual environment"
-	deactivate
-
+	CREATE_VIRTUALENV=1
 fi
 
+if [ $CREATE_VIRTUALENV == 1 ]; then
+	echo "Creating virtual enviroment '$VIRTUALENV'"
+	virtualenv $VIRTUALENV
+
+	if [ ! -d $VIRTUALENV ]; then
+		echo "Failed creating virtual environment!"
+		exit 1
+	else
+		echo "Virtual environment created. Starting..."
+		source $VIRTUALENV/bin/activate
+
+		pushd "$DIR"
+		python setup.py develop
+		popd
+
+		echo "Stopping virtual environment"
+		deactivate
+	fi
+fi
+
+
+DB_FOLDER="$DIR/../db"
+
+source $DB_FOLDER/config.sh
 
 WIPE_DB=0
-if [[ -z `sudo -u postgres psql -Atqc '\list $MOX_DB'` ]]; then
-	echo "Database $MOX_DB already exists in PostgreSQL"
-	read -p "Do you want to overwrite it? (y/n): " -n 1 -r
-	echo
-	if [[ $REPLY =~ ^[Yy]$ ]]; then
-		WIPE_DB=1
+
+if [ ! -z $ALWAYS_CONFIRM ]; then
+	WIPE_DB=1
+else
+	if [[ (! -z `command -v psql`) && (! -z `sudo -u postgres psql -Atqc "\list $MOX_DB"`) ]]; then
+		echo "Database $MOX_DB already exists in PostgreSQL"
+		read -p "Do you want to overwrite it? (y/n): " -n 1 -r
+		echo
+		if [[ $REPLY =~ ^[Yy]$ ]]; then
+			WIPE_DB=1
+		fi
+	else
+		echo "DB does not exist!"
 	fi
 fi
 
 if [ $WIPE_DB == 1 ]; then
 	# Install Database
 	source $VIRTUALENV/bin/activate
-	DB_FOLDER="$DIR/../db"
 
 	echo "Installing database"
